# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/infra/stack-pool/servers": {
      "get": {
        "operationId": "InfraStackPoolController_listServers",
        "summary": "Lister les VPS provisionnés avec pagination et filtres",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page à retourner",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Nombre de VPS par page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "contractState",
            "required": false,
            "in": "query",
            "description": "Filtrer selon létat contractuel du VPS",
            "schema": {
              "example": "active",
              "enum": [
                "active",
                "terminated"
              ],
              "type": "string"
            }
          },
          {
            "name": "provisioningStatus",
            "required": false,
            "in": "query",
            "description": "Filtrer selon le statut technique du provisioning",
            "schema": {
              "example": "READY",
              "enum": [
                "REQUESTED",
                "PROVISIONING",
                "READY",
                "DEGRADED",
                "RELEASE_REQUESTED",
                "RELEASED",
                "ERROR"
              ],
              "type": "string"
            }
          },
          {
            "name": "requestedFrom",
            "required": false,
            "in": "query",
            "description": "Filtrer les VPS demandés à partir de cette date ISO",
            "schema": {
              "example": "2026-03-28T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "requestedTo",
            "required": false,
            "in": "query",
            "description": "Filtrer les VPS demandés jusquà cette date ISO",
            "schema": {
              "example": "2026-03-31T23:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "infra-stack-pool"
        ]
      }
    },
...
    "/infra/stack-pool/provision": {
      "post": {
        "operationId": "InfraStackPoolController_provision",
        "summary": "Demander le provisionnement manuel de nouveaux VPS/stacks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionStackCapacityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow(s) GitHub dispatché(s)"
          }
        },
        "tags": [
          "infra-stack-pool"
        ]
      }
    },
...
    "/infra/stack-pool/release": {
      "post": {
        "operationId": "InfraStackPoolController_release",
        "summary": "Libérer une stack réservée ou résilier un VPS",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseStackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "infra-stack-pool"
        ]
      }
    },
...
    "/webhooks/whatsapp/connected": {
      "post": {
        "operationId": "WebhooksController_whatsappConnected",
        "summary": "Webhook appelé quand WhatsApp se connecte",
        "description": "Endpoint appelé par le whatsapp-connector quand le client WhatsApp est prêt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsAppConnectedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion traitée avec succès"
          }
        },
        "tags": [
          "webhooks"
        ]
      }
    },
    "/webhooks/whatsapp/events": {
      "post": {
        "operationId": "WebhooksController_whatsappEvents",
        "summary": "Webhook pour tous les events WhatsApp",
        "description": "Endpoint appelé par le whatsapp-connector pour tous les events (qr, authenticated, etc.)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsAppEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event traité avec succès"
          }
        },
        "tags": [
          "webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProvisionStackCapacityDto": {
        "type": "object",
        "properties": {
          "vpsCount": {
            "type": "number",
            "description": "Nombre de VPS a provisionner",
            "example": 2
          },
          "stacksPerVps": {
            "type": "number",
            "description": "Nombre de stacks a creer par VPS",
            "example": 2
          },
          "serverType": {
            "type": "string",
            "description": "Type de serveur Hetzner",
            "example": "CPX21"
          },
          "location": {
            "type": "string",
            "description": "Localisation Hetzner",
            "example": "fsn1"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Numéro utilisateur à prioriser pour l'allocation",
            "example": "+237612345678"
          },
          "pairingToken": {
            "type": "string",
            "description": "Pairing token frontend pour pousser les états WebSocket"
          },
          "deviceType": {
            "type": "string",
            "description": "Type d'appareil de la session en attente",
            "enum": [
              "mobile",
              "desktop"
            ]
          },
          "reason": {
            "type": "string",
            "description": "Raison métier ou opérationnelle",
            "example": "keep-minimum-free-stacks"
          }
        }
      },
      "ReleaseStackDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "agentId": {
            "type": "string"
          },
          "serverId": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "example": "manual-release"
          },
          "deleteServerWhenEmpty": {
            "type": "boolean",
            "description": "Supprimer aussi le VPS si plus aucune stack utile ne reste dessus",
            "default": false
          }
        }
      },
...
      "WhatsAppConnectedDto": {
        "type": "object",
        "properties": {}
      },
      "WhatsAppEventDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}